requirements.txt
# General
PyStaticConfiguration>=0.5.3

# Feed reading
feedparser>=5.1.3
pytz>=2013b

# Email sending
requests>=1.2.0

# Heroku API
heroku>=0.1.2

# Web
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
distribute==0.6.24
gunicorn==0.17.4
wsgiref==0.1.2

